pyproject.toml
[project]
name = "goal-driven-coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cerebras-cloud-sdk>=1.59.0",
    "openai>=2.7.2",
    "openai-agents>=0.5.0",
    "python-dotenv>=1.0.1",
]
